// apps/web/components/dynamic-ui-system/internal/types.d.ts
export type CondensedTShirtSizes = "s" | "m" | "l";
export type TShirtSizes = CondensedTShirtSizes | "xs" | "xl";
export type ResponsiveSpacingToken = TShirtSizes;
export type ShadowSize = TShirtSizes;
export type SpacingToken = StaticSpacingToken | ResponsiveSpacingToken;
export type opacity = 0 | 10 | 20 | 30 | 40 | 50 | 60 | 70 | 80 | 90 | 100;
export type ColorScheme =
  | "neutral"
  | "brand"
  | "accent"
  | "info"
  | "danger"
  | "warning"
  | "success";
export type ColorCategory = "on-solid" | "on-background";
export type ColorWeight = "weak" | "medium" | "strong";
export type Colors =
  | `${ColorScheme}-${ColorWeight}`
  | `${ColorScheme}-alpha-${ColorWeight}`;